Refactored typescript and added 'BaseClient.ts' class
